server/medapi/requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==1.11.5
django-json-widget==0.1.0
django-pdfkit==0.3.1
django-rest-swagger==2.1.2
djangorestframework==3.6.4
html5lib==0.999999999
idna==2.6
itypes==1.1.0
Jinja2==2.10
lxml==4.1.0
MarkupSafe==1.0
matplotlib==2.0.2
mimesis==0.0.10
numpy==1.13.3
olefile==0.44
openapi-codec==1.3.2
pandas==0.20.3
pdfkit==0.6.0
Pillow==4.3.0
pip-autoremove==0.9.0
psycopg2==2.7.3.1
Pygments==2.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.2
reportlab==3.4.0
requests==2.18.4
scikit-learn==0.19.0
scipy==0.19.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
transliterate==1.10
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.0.6
XlsxWriter==1.0.0
